scripts/requirements.txt
# Requirements for Hub admin scripts
# Install: pip install -r scripts/requirements.txt

openpyxl>=3.1.0        # Excel export
psycopg2-binary>=2.9.0 # PostgreSQL connection
python-dotenv>=1.0.0   # Environment loading
